Use size_t, not int, for word sizes for benchmarking code.

+++ src/bench/axHelm/benchmarkAx.cpp
@@ -18,7 +18,7 @@ struct CallParameters{
   bool constCoeff;
   bool poisson;
   bool computeGeom;
-  int wordSize;
+  size_t wordSize;
   int Ndim;
   bool stressForm;
   std::string suffix;
@@ -59,7 +59,7 @@ occa::kernel benchmarkAx(int Nelements,
                          bool constCoeff,
                          bool poisson,
                          bool computeGeom,
-                         int wordSize,
+                         size_t wordSize,
                          int Ndim,
                          bool stressForm,
                          int verbosity,
@@ -344,7 +344,7 @@ template occa::kernel benchmarkAx<int>(int Nelements,
                                        bool constCoeff,
                                        bool poisson,
                                        bool computeGeom,
-                                       int wordSize,
+                                       size_t wordSize,
                                        int Ndim,
                                        bool stressForm,
                                        int verbosity,
@@ -358,7 +358,7 @@ template occa::kernel benchmarkAx<double>(int Nelements,
                                           bool constCoeff,
                                           bool poisson,
                                           bool computeGeom,
-                                          int wordSize,
+                                          size_t wordSize,
                                           int Ndim,
                                           bool stressForm,
                                           int verbosity,
